Add queueQuery method to Connection
This adds a queueQuery method to Connection that allows you to queue up
multiple queries which will be sent all at once when any of the returned
wait handles are joined. This is intended for use in places where many small
queries are executed and the overhead of opening up many connections to
service them asynchronously is too much.



git-svn-id: file:///var/svn/beatbox/trunk@171 4482dde9-802c-4b55-bdde-2e54f8a989d0