Many changes
Most significant change is reworking the initialisation code, you now pass
a path to a configuration file to an initialisation function. This does the
setup, instead of when the `init.php` file is included.

The autloader now only uses `HH\set_autoload_paths` with some code to help in
error conditions.

The classmap builder has been re-written (again). It is now more library-ish
and can be easily called from other sources. It tries to produce a map that is
as flexible as possible.

XHP and PEAR have been moved into a lib/ subfolder, this is just so all our
"third-party" code is a single place instead of cluttering up the top-level
directory.

Adds a work-in-progress code generation module to help with generating
PHP/Hack code.


git-svn-id: file:///var/svn/beatbox/trunk@188 4482dde9-802c-4b55-bdde-2e54f8a989d0